requirements.txt
discord.py
#discordlevelingcard
python-dotenv
requests
mee6_py_api